requirements.txt
# ============================================================
# UFM Admission Bot — Python Dependencies
# ============================================================

# --- Web Framework ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# --- Database (PostgreSQL + pgvector) ---
psycopg2-binary>=2.9.9

# --- AI / ML ---
numpy>=1.24.0

# --- Data Validation ---
pydantic>=2.5.0

# --- Config ---
pyyaml>=6.0.1
python-dotenv>=1.0.0

# --- Template Engine (PromptManager dùng Jinja2 render user_prompt) ---
Jinja2>=3.1.2

# --- HTTP Client ---
httpx>=0.25.0

# --- Auth & Security ---
PyJWT>=2.8.0

# --- FastAPI Form/File Upload (Admin router dùng Form(), File()) ---
python-multipart>=0.0.6
